refactor(css): move CSS files to centralized directory
Moved all CSS files from individual component and page directories to src/css/
Updated import statements in JSX files to reflect new paths (e.g., './RegisterForm.css' to '../css/RegisterForm.css')
This improves project organization and maintainability by centralizing styles in one location
Minor updates to package-lock.json from dependency installation
