+++ src/pages/HomePage.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import "./HomePage.css";
+import "../css/HomePage.css";
 
 export default function HomePage() {
   return (
