+++ src/pages/RegisterPage.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import "./RegisterPage.css";
+import "../css/RegisterPage.css";
 
 export default function RegisterPage() {
   const navigate = useNavigate();
